% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-fhm.R
\docType{package}
\name{fhm-package}
\alias{fhm}
\alias{fhm-package}
\title{fhm}
\description{
\code{flat_hash_map} provides a drop-in replacement for \code{std::unodered_map}
and \code{std::unordred_set}. The hash map provided in \code{flat_hash_map} considerably
outperforms \code{std::unordered_map} in a number of different benchmarks. 
For a detailed write-up on the methods and some benchmarks, see this blogpost:
\url{https://probablydance.com/2017/02/26/i-wrote-the-fastest-hashtable/}.
}
\details{
This package makes the \code{flat_hash_map} methods available to other R 
packages, via header files.

The original source code was written by Malte Skarupke.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skarupke/flat_hash_map}
  \item \url{https://github.com/ChrisMuir/r-fhm}
  \item Report bugs at \url{https://github.com/ChrisMuir/r-fhm/issues}
}

}
\author{
\strong{Maintainer}: Chris Muir \email{chrismuirrva@gmail.com}

Authors:
\itemize{
  \item Malte Skarupke [copyright holder]
}

}
